refactoring TGS_inference
